requirements.txt
Flask~=2.2.2
python-dotenv~=0.21.0
requests~=2.28.1
redis~=4.3.4
swagger-ui-py~=22.7.13
psycopg2~=2.9.5
minio~=7.1.12
